requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# ── AI ────────────────────────────────────────────────────────────────────────
openai>=1.30.0

# ── Database ──────────────────────────────────────────────────────────────────
supabase>=2.3.0

# ── Image rendering ───────────────────────────────────────────────────────────
playwright>=1.44.0
Pillow>=10.3.0        # Windows dev fallback only

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0         # Required by FastAPI TestClient
